Fix pipeline editor.
Fetching the connector data on-demand while adding a node
breaks the layout as we only know the size once it's too late
(e.g., position has been determined already).
We do know all the ConnectorDescr data when adding a new one
so we just generate it instead of fetching it on-demand.